requirements.txt
boto3==1.20.34
fastapi==0.71.0
gunicorn==20.1.0
openpyxl==3.0.9
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
tabula-py==2.3.0
uvicorn==0.16.0
moviepy==1.0.3
PyYAML==6.0.0
pyPDF2==1.26.0
pandas_read_xml==0.3.1
ruamel.yaml==0.17.21
img2pdf==0.4.4
aspose.words==22.9.0
aspose-diagram==22.9.0
pillow-heif==0.7.0
JPype1==1.4.0
ffmpeg==1.4
pre-commit==2.20.0
flake8
black
psd-tools==1.9.22
Faker==15.3.2
nltk==3.7
yake==0.4.8
wheel==0.38.4
spacy==3.4.3
textstat==0.7.3
readability==0.3.1
svglib==1.4.1
regex==2022.10.31
reportlab==3.6.12
spacy-legacy==3.0.10
spacy-loggers==1.0.3
selenium==4.6.1
webdriver-manager==3.8.5
fastapi-utils==0.2.1
opencv-python==4.6.0.66
qrcode==7.3.1
alembic==1.8.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
Werkzeug==2.2.2
fastapi-jwt-auth==0.5.0
phonenumbers==8.13.2